+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/Components/SpatialPingComponent.h
@@ -9,6 +9,32 @@
 
 DECLARE_LOG_CATEGORY_EXTERN(LogSpatialPingComponent, Log, All);
 
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnRecordPing, float, Ping);
+
+USTRUCT(BlueprintType)
+struct FSpatialPingAverageData
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	float LastMeasurementsWindowAvg;
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	float LastMeasurementsWindowMin;
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	float LastMeasurementsWindowMax;
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	int WindowSize;
+
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	float TotalAvg;
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	float TotalMin;
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	float TotalMax;
+	UPROPERTY(BlueprintReadWrite, Category = SpatialPing)
+	int TotalNum;
+};
+
 /*
  Offers a configurable means of measuring round-trip latency in SpatialOS deployments.
  This component should be attached to a player controller.
@@ -31,6 +57,10 @@ public:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SpatialPing)
 	float TimeoutLimit = 4.0f;
 
+	// The number of ping measurements recorded for the rolling average.
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SpatialPing)
+	int PingMeasurementsWindowSize = 20;
+
 	virtual void BeginPlay() override;
 	virtual void EndPlay(const EEndPlayReason::Type EndPlayReason) override;
 
@@ -45,6 +75,14 @@ public:
 	// Returns the latest raw round-trip ping value in seconds.
 	UFUNCTION(BlueprintCallable, Category = "SpatialGDK|Ping")
 	float GetPing() const;
+
+	// Returns the average, min, and max values for the last PingMeasurementsWindowSize measurements as well as the total average, min, and max.
+	UFUNCTION(BlueprintCallable, Category = "SpatialGDK|Ping")
+	FSpatialPingAverageData GetAverageData() const;
+
+	// Multicast delegate that will be broadcast whenever a new ping measurement is recorded.
+	UPROPERTY(BlueprintAssignable, Category = "SpatialGDK|Ping")
+	FOnRecordPing OnRecordPing;
 
 private:
 	float RoundTripPing;
@@ -79,4 +117,13 @@ private:
 
 	UFUNCTION(Server, Unreliable, WithValidation)
 	virtual void SendServerWorkerPingID(uint16 PingID);
+
+	void RecordPing(float Ping);
+
+	TArray<float> LastPingMeasurements;
+
+	float TotalPing = 0.0f;
+	float TotalMin = 1.0f;
+	float TotalMax = 0.0f;
+	int TotalNum = 0;
 };